+++ src/components/dialog/LanguageDialog.tsx
@@ -1,39 +1,17 @@
-import type { ReactNode, PropsWithChildren } from 'react'
+import type { PropsWithChildren, ReactNode } from 'react'
 import type { DialogProps } from '@/src/components/dialog/Dialog'
 import { Dialog } from '@/src/components/dialog/Dialog'
-import type { Language } from '@/src/localization/util'
-import { LanguageUtil } from '@/src/localization/util'
-import type { PropsForTranslation, Translation } from '@/src/localization/useTranslation'
-import { useTranslation } from '@/src/localization/useTranslation'
-import { useLanguage } from '@/src/localization/LanguageProvider'
+import { LocalizationUtil } from '@/src/i18n/util'
+import { useLocale } from '@/src/i18n/LocaleProvider'
 import { Select, SelectOption } from '@/src/components/user-action/select/Select'
 import { SolidButton } from '@/src/components/user-action/Button'
+import { useStandardTranslation } from '@/src/i18n/useTranslation'
+import type { SupportedLocale } from '@/src/i18n/translations'
 
-type LanguageDialogTranslation = {
-  language: string,
-  chooseLanguage: string,
-  done: string,
-} & Record<Language, string>
-
-const defaultLanguageDialogTranslation: Translation<LanguageDialogTranslation> = {
-  en: {
-    language: 'Language',
-    chooseLanguage: 'Choose your language',
-    done: 'Done',
-    ...LanguageUtil.languagesLocalNames
-  },
-  de: {
-    language: 'Sprache',
-    chooseLanguage: 'Wähle deine bevorzugte Sprache',
-    done: 'Fertig',
-    ...LanguageUtil.languagesLocalNames
-  }
-}
-
-type LanguageDialogProps = Omit<DialogProps, 'titleElement' | 'description'> & {
+type LanguageDialogProps = Omit<DialogProps, 'titleElement' | 'description'> & PropsWithChildren<{
   titleOverwrite?: ReactNode,
   descriptionOverwrite?: ReactNode,
-}
+}>
 
 /**
  * A Dialog for selecting the Language
@@ -41,14 +19,13 @@ type LanguageDialogProps = Omit<DialogProps, 'titleElement' | 'description'> & {
  * The State of open needs to be managed by the parent
  */
 export const LanguageDialog = ({
-                                 overwriteTranslation,
                                  onClose,
                                  titleOverwrite,
                                  descriptionOverwrite,
                                  ...props
-                               }: PropsForTranslation<LanguageDialogTranslation, PropsWithChildren<LanguageDialogProps>>) => {
-  const { language, setLanguage } = useLanguage()
-  const translation = useTranslation([defaultLanguageDialogTranslation], overwriteTranslation)
+                               }: LanguageDialogProps) => {
+  const { locale, setLocale } = useLocale()
+  const translation = useStandardTranslation()
 
   return (
     <Dialog
@@ -59,15 +36,15 @@ export const LanguageDialog = ({
     >
       <div className="w-64">
         <Select
-          value={language}
-          onValueChanged={(language: string) => setLanguage(language as Language)}
+          value={locale}
+          onValueChanged={(language: string) => setLocale(language as SupportedLocale)}
           contentPanelProps={{ containerClassName: 'z-200' }}
           buttonProps={{
-            selectedDisplay: value => translation(value as Language)
+            selectedDisplay: locale => LocalizationUtil.languagesLocalNames[locale]
           }}
         >
-          {LanguageUtil.languages.map((language) => (
-            <SelectOption key={language} value={language}>{translation(language)}</SelectOption>
+          {LocalizationUtil.locals.map((local) => (
+            <SelectOption key={local} value={local}>{LocalizationUtil.languagesLocalNames[local]}</SelectOption>
           ))}
         </Select>
         <div className="flex-row-4 mt-3 justify-end">