+++ src/components/dialog/ThemeDialog.tsx
@@ -3,14 +3,11 @@ import type { DialogProps } from '@/src/components/dialog/Dialog'
 import { Dialog } from '@/src/components/dialog/Dialog'
 import { MonitorCog, MoonIcon, SunIcon } from 'lucide-react'
 import clsx from 'clsx'
-import type { ThemeType, ThemeTypeTranslation } from '@/src/theming/useTheme'
+import type { ThemeType } from '@/src/theming/useTheme'
 import { ThemeUtil, useTheme } from '@/src/theming/useTheme'
-import type { FormTranslationType } from '@/src/localization/defaults/form'
-import { formTranslation } from '@/src/localization/defaults/form'
-import type { PropsForTranslation, Translation } from '@/src/localization/useTranslation'
-import { useTranslation } from '@/src/localization/useTranslation'
 import { Select, SelectOption } from '@/src/components/user-action/select/Select'
 import { SolidButton } from '@/src/components/user-action/Button'
+import { useStandardTranslation } from '@/src/i18n/useTranslation'
 
 type ThemeIconProps = {
   theme: ThemeType,
@@ -32,21 +29,6 @@ const ThemeIcon = ({ theme, className }: ThemeIconProps) => {
   }
 }
 
-type ThemeDialogTranslationAddon = {
-  chooseTheme: string,
-}
-
-type ThemeDialogTranslation = ThemeDialogTranslationAddon & ThemeTypeTranslation & FormTranslationType
-
-const defaultConfirmDialogTranslation: Translation<ThemeDialogTranslationAddon> = {
-  en: {
-    chooseTheme: 'Choose your preferred color theme.',
-  },
-  de: {
-    chooseTheme: 'Wähle dein bevorzugtes Farbschema.',
-  }
-}
-
 type ThemeDialogProps = Omit<DialogProps, 'titleElement' | 'description'> & {
   titleOverwrite?: ReactNode,
   descriptionOverwrite?: ReactNode,
@@ -58,18 +40,17 @@ type ThemeDialogProps = Omit<DialogProps, 'titleElement' | 'description'> & {
  * The State of open needs to be managed by the parent
  */
 export const ThemeDialog = ({
-                              overwriteTranslation,
                               onClose,
                               titleOverwrite,
                               descriptionOverwrite,
                               ...props
-                            }: PropsForTranslation<ThemeDialogTranslation, PropsWithChildren<ThemeDialogProps>>) => {
+                            }: PropsWithChildren<ThemeDialogProps>) => {
   const { theme, setTheme } = useTheme()
-  const translation = useTranslation([defaultConfirmDialogTranslation, formTranslation, ThemeUtil.translation], overwriteTranslation)
+  const translation = useStandardTranslation()
 
   return (
     <Dialog
-      titleElement={titleOverwrite ?? translation('theme')}
+      titleElement={titleOverwrite ?? translation('themes', { count: 1 })}
       description={descriptionOverwrite ?? translation('chooseTheme')}
       onClose={onClose}
       {...props}
@@ -86,7 +67,7 @@ export const ThemeDialog = ({
             selectedDisplay: (value) => (
               <div className="flex-row-2 items-center">
                 <ThemeIcon theme={theme}/>
-                {translation(value as ThemeType)}
+                {translation('themeMode', { theme: value })}
               </div>
             ),
             className: 'min-w-32',
@@ -96,7 +77,7 @@ export const ThemeDialog = ({
             <SelectOption key={theme} value={theme} className="gap-x-6 justify-between">
               <div className="flex-row-2 items-center">
                 <ThemeIcon theme={theme}/>
-                {translation(theme as ThemeType)}
+                {translation('themeMode', { theme: theme })}
               </div>
             </SelectOption>
           ))}