+++ src/components/user-action/input/InsideLabelInput.tsx
@@ -47,7 +47,8 @@ export const InsideLabelInput = forwardRef<HTMLInputElement, InsideLabelInputPro
         aria-hidden={true}
         data-display={isFocused || !!value ? 'small' : 'full'}
         className={clsx(
-          'absolute left-4 top-2 transition-all pointer-events-none touch-none',
+          // margin left to account for the border which is ignored for absolute positions
+          'absolute left-4 ml-0.5 top-2 transition-all delay-25 pointer-events-none touch-none',
           'data-[display=small]:top-2 data-[display=small]:h-force-4.5 data-[display=small]:typography-caption-sm data-[display=small]:overflow-y-hidden',
           'data-[display=full]:top-1/2 data-[display=full]:-translate-y-1/2 data-[display=full]:typography-body-md'
         )}